requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Date/Time
python-dateutil==2.8.2

# AI/ML
torch==2.1.0
transformers==4.35.0
huggingface-hub==0.19.0

# Image Generation
openai==1.3.0
Pillow==10.1.0

# TTS
google-cloud-texttospeech==2.14.1

# Video Generation
moviepy==1.0.3
numpy==1.24.3
